Examples/mobc: apply rename migration(IfWrapper -> hal)
