+++ Examples/mobc/src/src_user/Drivers/Aocs/aobc.h
@@ -5,7 +5,7 @@
 #ifndef AOBC_H_
 #define AOBC_H_
 
-#include <src_core/IfWrapper/uart.h>
+#include <src_core/hal/uart.h>
 #include <src_core/Drivers/Super/driver_super.h>
 #include <src_core/System/TimeManager/obc_time.h>
 #include <src_core/TlmCmd/common_cmd_packet.h>
