+++ Examples/mobc/src/src_user/Drivers/Com/gs.c
@@ -9,7 +9,7 @@
 #include <string.h>
 #include <stdint.h>
 
-#include "../../IfWrapper/ccsds_user.h"
+#include "../../hal/ccsds_user.h"
 #include "../../TlmCmd/Ccsds/tc_transfer_frame.h"
 #include <src_core/Drivers/Super/driver_super.h>
 #include <src_core/TlmCmd/packet_handler.h>
