+++ Examples/mobc/src/src_user/Script/AddSection/settings.pl
@@ -18,7 +18,7 @@ sub GetSetting {
 	$SETTING{'SEARCH_PATH'} = [
 		'Applications',
 		'Drivers',
-		'IfWrapper',
+		'hal',
 		'Library',
 		'Settings',
 		'TlmCmd'
