+++ Examples/mobc/src/src_user/hal/if_list.h
@@ -2,8 +2,8 @@
 #define IF_LIST_H_
 
 
-#include <src_core/IfWrapper/ccsds.h>
-#include <src_core/IfWrapper/uart.h>
+#include <src_core/hal/ccsds.h>
+#include <src_core/hal/uart.h>
 
 typedef enum
 {
